refactor: extract reviewer_ui.py from __init__.py
Move shortcut functions, reviewer bottom bar customization, and
Reviewer monkey-patching (~95 lines) into reviewer_ui.py.
setup_reviewer_ui() handles all the wiring in one call.
